+++ packages/helpers/src/customPropTypes.js
@@ -57,3 +57,5 @@ export const FocusPropTypes = {
 
 export const RefOf = (propType = PropTypes.object) =>
   PropTypes.oneOfType([PropTypes.func, PropTypes.shape({ current: propType })]);
+
+export const DOMElementType = PropTypes.oneOfType([PropTypes.instanceOf(Element), PropTypes.instanceOf(ShadowRoot)]);