+++ packages/Modal/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.22-next.0
+
+### Patch Changes
+
+- Updated dependencies [63f9a05]
+  - @paprika/overlay@1.0.13-next.0
+  - @paprika/helpers@2.2.1-next.0
+  - @paprika/button@1.1.11-next.0
+
 ## 1.0.21
 
 ### Patch Changes