+++ packages/Table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.2.20-next.0
+
+### Patch Changes
+
+- Updated dependencies [63f9a05]
+  - @paprika/helpers@2.2.1-next.0
+  - @paprika/button@1.1.11-next.0
+  - @paprika/overflow-menu@2.0.5-next.0
+
 ## 1.2.19
 
 ### Patch Changes